Improved Validation Accuracy
Realized that the the validation code for siglip_vision wasn't using
padding="max_length". This had a big result on the accuracies. In fact,
it complete inverted the ordering of the three different prompt variations.
The simplest "A photo of a {text description}" gives the best performance.

So updated the documentation and the code to use that prompt template in
both the siglip_vision and siglip_text validation code.