Renamed model_repository & changed default language_id_threshold
model_repository -> model-repository; done in preparation for merging main back into this branch

During validation, I was seeing significant drop in chF2++ scores when the src_lang was not
given. This is because language identification performs worse on some languages. It then
was doing language detection on every sentence which caused it to error out sometimes. This
affected the score. Typically, it identifies a language like 'bod' (should be cmn) and then
that language is not supported by seamlessM4T